mymap/requirements.txt
asgiref==3.6.0
Django==4.1.8
django-cors-headers==4.0.0
django-filter==22.1
djangorecipe==2.2.1
djangorestframework==3.14.0
djangorestframework-gis==1.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
zc.buildout==3.0.1
zc.recipe.egg==2.0.7
